Move configuration defaults to parse code
This way the configuration parser test can test for the default values which is
the simplest place for these tests. It does create a somewhat weird
relationship with `DefaultConfig` in the test code, essentially still testing
the defaulting logic. Other tests can avoid that by not depending on the
default logic in the tests.