Finalize integration test
Now it's something that exits with code 0 on success and code 1 on failure. In
theory this makes it usable in a CI pipeline but in practice it hasn't been
tested in a CI environment and I'm not sure how the k3d CLI would behave in
one. Maybe it'd work in general because it should only really depend on Docker
but I'm especially suspicious of the registry setup because it binds a port on
the host that runs Docker which in CI hosts likely isn't the agent which itself
is also running in Docker in Kubernetes.