Add integration test to Jenkinsfile
Locally the integration test runs on k3d (k3s in Docker). However, I couldn't
get k3s running in a pod in a kubernetes-in-kubernetes manner. So this CI
integration test uses kind (kuberentes-in-docker) instead. Kind is a project
similar to k3d, developed by the Kubernetes project itself.

The specific Pod template is based on a [Running KIND inside A Kubernetes
Cluster for Continuous Integration][1] blog post.

In theory, the Docker image built earlier in the job could be loaded into the
Docker daemon running in the `kind-cluster` container. However, I'm not exactly
sure how to do that so currently I have avoided this problem by pushing the
Docker image onto Docker Hub before running the integration tests.

[1]: https://d2iq.com/blog/running-kind-inside-a-kubernetes-cluster-for-continuous-integration