Start kind container in root dir
Jenkins changes the working directory to be the same for all containers but
this image expects to start in the root folder. Work around this by explicitly
starting the entrypoint in the root folder.

The last line of `entrypoint.sh` is `exec "$@"`. Previously `/bin/bash` was
given as `args`. To keep the start command effectively the same, `/bin/bash` is
provided as an argument directly to `/entrypoint.sh`.